Two class projects
PH503_MUON_15 /

I have studied the origins of the anomaly and briefly discussed the methodology employed at Fermilab to measure the muon’s magnetic anomaly, and explored the implications of this deviation. This was a class project for the course on Atomic and Molecular physics (PH503) taken by Prof. Bipul Bhyuan at IIT Guwahati. 

anamalous_a /

I specifically calculate the one-loop correction to the magnetic moment, famously known as Schwinger’s term, which contributes an amount of $\alpha/\pi$ to the anomaly. The anomalous magnetic moment, defined as $a_\mu = \frac{g - 2}{2}$, quantifies this deviation from the Dirac prediction of g equal to 2. This was a class project for the course on Particle Physics (PH544) taken by Dr. Debasish Borah at IIT Guwahati.